requirements.txt
asgiref==3.3.1
Django==3.1.3
django-vanilla-views==2.0.0
Pillow==8.0.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
